Mk3 integration (#2)
* removing sudo for normal build

* Fix missing (hex) in D5 code

* feutre -> furtif

Translation error of Stealth

* Typo in some hex default values

* Fix DEBUG_BUILD

* Fix case sensitive include

* Code cleanup

* EEPROM_BACKLIGHT_TIMEOUT min is 1 sec

* Version to 1.0.1

* Move few D-codes command from Marlin_main to Dcodes
D80, D81 and D106 have been added to `Marlin_main.cpp` but all other D-codes moved to `Dcodes.cpp/.h`

* Reset all stats during Shipping prep (#2563)

* Reset all stats during Shipping prep
Last print stats were missing

* keep old indentation

Co-authored-by: D.R.racer <drracer@drracer.eu>

* First attempt for Z probe checking during selftest on the MK3/S

* Raise on nozzle crash

* Fix go up

* Make ALL D-codes available for DEBUG mode
until now D3 and D5 needed to be defined separately

* Forgot to modify Marlin_main

* :recycle:Code optimization

* Fix missing (hex) in D5 code (#2560)

* Fix missing (hex) in D5 code

* Make ALL D-codes available for DEBUG mode
until now D3 and D5 needed to be defined separately

* Forgot to modify Marlin_main

* :art:Change CRLF line ending to LF

* Bugfix for last PR
I forgot to change also the
- `#ifdef DEBUG_DCODE_3` to `#if defined DEBUG_DCODE3 || defined DEBUG_DCODES`
- `#ifdef DEBUG_DCODE_5` to `#if defined DEBUG_DCODE5 || defined DEBUG_DCODES`
in the `Dcodes.h` file which I added to `Dcodes.cpp`.

Due to this issue the "Debug" version fails during compiling.

Sorry for that.

* Cleanup useless `break;`

* Remove empty line

* Always reset e_steps between blocks

If e_steps are scheduled, but not ticked, they're just lost.
Only carry over the pressure state.

* Remove several globals by using a single target pressure

In the current code we initialize the LA state on-demand already at the
right step, which makes keeping track of the tick position no longer
necessary.

Make the advance ISR almost stateless by removing the last vestiges of
the original implementation and introduce a single target pressure. This
will be needed later in order to trigger the LA isr inside the cruising
phase.

* Release excess pressure within cruising blocks

LA assumes all the nozzle pressure is released at the end of each
extrusion, which makes calculating the required pressure advance during
travels and retracts not normally necessary.

This is not always true in our planner, since the E axis is explicitly
ignored when not in use, but also due to E-jerk allowing a non-linear
jump in speed. And since the compression factor is currently tied by XYZ
axes and not independently calculated, this can result in a wrong
estimation of final pressure in several conditions.

To avoid overburdening the planner, change the underlying assumptions
about backpressure:

1) Pressure is no longer lost when LA is disabled: if a retract is
followed by an unretract of the same length, the pressure will be likely
maintained entirely. This also holds true during travels, as long as the
retract length can overcome all the backpressure (which is the case in
all but the most noodly materials)

2) Pressure is released as soon as possible during travels: we now
enable LA also during travels, but under the sole condition of undoing
excess pressure.

We do that by checking for backpressure at the start of any segment
without an acceleration phase that doesn't have any E-steps (a result
which can happen due to the above). If pressure is not nominal, we run
the extruder in reverse at maximum jerk as long as the segment allows
us, since proper acceleration would be prohibitive at this stage. As the
pressure difference resulting by the above is still _very_ low, any wipe
or short travel will be able to equalize the nozzle pressure *before*
extrusion is resumed, avoiding ooze.

* :bug:Flip lcd_detect_IRsensor logic

* Do not overflow during LA acceleration limiting

Perform the check one step earlier, avoiding 32bit overflow for very low
compression factors.

Fixes #2566 (although for K15 to have effect the conversion probably
needs to be adjusted on the low end)

* Unify LA for all trapezoid steps

Handle uniformly compression & decompression at any stage of the
trapezoid.

Compared to before, this now enables LA compression also in the cruising
step (handling the converse of a chained wipe), as well as decompression
during acceleration.

Both of these can happen as a result of jerk moves, but are incredibly
rare. This is mostly needed to allow rapid decompression directly at the
acceleration step during travels between a retraction&deretraction.

We also check for the pressure level in a single place, reducing code
size as well as disabling LA earlier when not needed for the rest of the
block.

* Increase the LA10->15 response

* Merge from main branch

* Auto stash before merge of "MK3" and "prusa-origin/MK3"

* merge

Co-authored-by: Thelvaen <thelvaen.mandel@gmail.com>
Co-authored-by: 3d-gussner <3d.gussner@gmail.com>
Co-authored-by: valère didon <vdidon@live.fr>
Co-authored-by: Alex Voinea <voinea.dragos.alexandru@gmail.com>
Co-authored-by: DRracer <drracer@seznam.cz>
Co-authored-by: D.R.racer <drracer@drracer.eu>
Co-authored-by: Yuri D'Elia <wavexx@thregr.org>